+++ packages/banana-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @banana/banana-react
 
+## 1.5.5
+
+### Patch Changes
+
+- Added gap parameter to carousel.
+- Updated dependencies
+  - banana-ui@1.5.5
+
 ## 1.5.4
 
 ### Patch Changes